chore: Split metactl into lib meta/control and bin metactl
