+++ src/meta/control/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "databend-common-meta-control"
+description = "common meta control"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+doctest = false
+test = true
+
+[dependencies]
+databend-common-meta-types = { workspace = true }
+databend-common-tracing = { workspace = true }
+databend-common-meta-client = { workspace = true }
+databend-common-meta-raft-store = { workspace = true }
+databend-common-meta-sled-store = { workspace = true }
+databend-meta = { workspace = true }
+
+futures = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+raft-log = { workspace = true }
+reqwest = { workspace = true }
+tokio = { workspace = true }
+tokio-stream = { workspace = true }
+url = { workspace = true }
+anyhow = { workspace = true }
+clap = { workspace = true }
+
+[dev-dependencies]
+anyhow = { workspace = true }
+
+[build-dependencies]
+databend-common-building = { workspace = true }
+
+[lints]
+workspace = true
