+++ src/meta/control/src/lib.rs
@@ -1,0 +1,24 @@
+// Copyright 2025 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#![allow(clippy::uninlined_format_args)]
+
+pub mod admin;
+pub mod args;
+pub mod export_from_disk;
+pub mod export_from_grpc;
+pub mod import;
+mod import_v004;
+pub(crate) mod reading;
+mod upgrade;
